requirements.txt
Django==1.9.8
argparse==1.2.1
dj-database-url==0.4.1
django-cors-headers==1.1.0
djangorestframework==3.4.5
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
wsgiref==0.1.2
